Extract MultiRecordSelect from MultipleObjectRecordSelect
